+++ service/src/main/java/com/authentication/service/impl/UserServiceImpl.java
@@ -9,18 +9,28 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.authentication.model.User;
+import com.authentication.repository.IAttendanceRepository;
 import com.authentication.repository.IUserRepository;
 import com.authentication.service.IUserService;
 import com.authentication.service.dto.UserDto;
+import com.authentication.service.exception.MethodNotAllowedException;
 import com.authentication.service.exception.ResourceNotFoundException;
 
 @Service
 public class UserServiceImpl implements IUserService {
 
-    @Autowired
-    private IUserRepository userRepository;
+    private final IUserRepository userRepository;
 
-    private ModelMapper modelMapper = new ModelMapper();
+    private final IAttendanceRepository attendanceRepository;
+
+    private final ModelMapper modelMapper = new ModelMapper();
+
+    @Autowired
+    public UserServiceImpl(IUserRepository userRepository, IAttendanceRepository attendanceRepository) {
+        super();
+        this.userRepository = userRepository;
+        this.attendanceRepository = attendanceRepository;
+    }
 
     @Override
     public List<UserDto> findAll() {
@@ -43,7 +53,7 @@ public class UserServiceImpl implements IUserService {
     }
 
     @Override
-    public UserDto updateUser(Long userId, UserDto userDto) throws ResourceNotFoundException {
+    public UserDto update(Long userId, UserDto userDto) throws ResourceNotFoundException {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found on :: " + userId));
         user.setFirstName(userDto.getFirstName());
@@ -55,12 +65,14 @@ public class UserServiceImpl implements IUserService {
     }
 
     @Override
-    public void deleteById(Long userId) throws ResourceNotFoundException {
-        Optional.of(userRepository.existsById(userId))
-                .orElseThrow(() -> new ResourceNotFoundException("User not found on :: " + userId));
-
+    public void deleteById(Long userId) throws ResourceNotFoundException, MethodNotAllowedException {
+        if (!userRepository.existsById(userId)) {
+            throw new ResourceNotFoundException("User not found on :: " + userId);
+        }
+        if (attendanceRepository.existsByUserId(userId)) {
+            throw new MethodNotAllowedException("Attendance associated on :: " + userId);
+        }
         userRepository.deleteById(userId);
-
     }
 
 }
